new libft

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: pgerbaud <pgerbaud@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/09 10:42:21 by pgerbaud          #+#    #+#             */
-/*   Updated: 2016/11/09 14:00:56 by pgerbaud         ###   ########.fr       */
+/*   Updated: 2016/11/17 17:19:55 by pgerbaud         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memcpy(void *restrict dst, const void *restrict src, size_t n)
 {
-	int	index;
+	size_t	index;
 
 	index = 0;
 	while (index < n)
